requirement.txt
altgraph==0.17.4
click==8.1.8
-e git+https://github.com/LAQ27504/Personal-CLI-python.git@56885af61b639b8ae4b1119b258288595dd39548#egg=laq_custom_cli_tools
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
powerline-status==2.7
prompt-toolkit==1.0.14
Pygments==2.19.1
PyInquirer==1.0.3
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyYAML==6.0.2
regex==2024.11.6
rich==13.9.4
shellingham==1.5.4
six==1.17.0
typer==0.15.1
typing_extensions==4.12.2
wcwidth==0.2.13
